requirements.txt
certifi==2018.8.24
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
psycopg2-binary==2.7.5
pytz==2018.5
uWSGI==2.0.17.1
